PreDeathChestEvent (#4)
* PreDeathChestEvent

Add an event that is called each time a player dies.

* Death Message Class

Move the data for the death message over to its own class file, utilising the event generated by the plugin instead of a direct call.

* Better Method

Instead of utilising separate methods to retrieve the item and to remove it, use a single method that does both.